Update workflow names and add reporting steps
The names for the API test workflows running with httpYac have been altered to include the reporters' names for greater clarity. Additionally, steps to publish test reports and results have been added in the "api-tests-assertions.yaml" workflow. This step allows for an immediate view of the testing result status and easier troubleshooting if necessary.